# backend/public/swagger.yaml
openapi: 3.0.0
info:
  title: Order Creation API (Fantastic Five)
  description: SENG2021 Order Creation API. To interact with our API, signup and login to generate a session token and then use this token to authroize use of the API. All endpoints with the lock icon will require this authentication to be accessed. Our API provides support for orders and invoices. We also have an email service for account/order creation, updating and order deletion.
  version: 2.0.0
servers:
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /v1/invoices/validate/{invoiceId}:
    put:
      summary: Validate an invoice by its ID
      security:
        - bearerAuth: []
      description: Validates the invoice specified by the invoiceId path parameter using an external API.
      tags:
        - Invoices
      parameters:
        - name: invoiceId
          in: path
          description: The ID of the invoice to validate
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          description: The JWT token required for updating the order
          schema:
            type: string
          example: "Bearer <token>"
      responses:
        '200':
          description: Invoice successfully validated
          content:
            application/json:
              schema:
                type: object
                example:
                  message: "Invoice successfully validated!"

        '400':
          description: Bad request if the invoice ID is missing or malformed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to validate invoice"
...